% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias_machine.R
\name{alias_machine}
\alias{alias_machine}
\title{Aliased effects}
\usage{
alias_machine(num_of_vars, l)
}
\arguments{
\item{num_of_vars}{A natural number. This tells the function how many variables are intended for your design. The limit is 26.}

\item{l}{This is the system of equations used to generate the aliasing structure. It is to have each row be one "cut" of the factorial design and must have the same number of coefficents (which take the values 0, 1, 2) as x.}
}
\value{
A matrix. Each row will be an estimable effect and all it's aliases.
}
\description{
During a fractional factorial experiment certain effects are fully confounded. This function will tell display the aliased effects for any effect that is still estimable
}
\examples{
\dontrun{
l <- matrix(c(1,1,0,0,0,1,2,1), nrow = 2, byrow = TRUE)
l
alias_machine(4, l)
}
}
