% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_rotation.r
\name{opt_rotation}
\alias{opt_rotation}
\title{Finds an Optimal Rotation Vector}
\usage{
opt_rotation(design)
}
\arguments{
\item{design}{This needs to be a coded design matrix using (0,1,2) where each row corresponds to one treatment.}
}
\value{
It will produce a list
\item{Rotation Vector}{The first element of the list will be the optimum rotation vector and a small message whether Resoluiton IV is obtainable or if this is minimal abberration}
\item{psuedo-design matrix}{This is the fake design matrix which is the coded effects of length 3 that are aliased with the intercept in the original design. This will not neccessarly be a regular fractional factorial design (hence psuedo)}
}
\description{
This function will produce an optimal rotation vector for some design. It will produce Resolution IV if possible and if not will find a minimal aberration. If multiple minimal aberration designs exist it will randomly choose one and ignore any "desirability" between factors.
}
\examples{
\donttest{
x <- c("a", "b", "c", "d", "e", "f")
l <- matrix(c(1,2,0,1,1,1, 0,1,1,2,0,0, 1,1,1,2,2,0), nrow = 3, byrow = TRUE)
trees <- c(0,0,0)
m <- TripleFrac:::part(x, l, trees)[[1]]
head(m)
what_frac(m)
opt_rotation(m)
}
}
