% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/what_frac.R
\name{what_frac}
\alias{what_frac}
\title{Alias Structure Identifier
Returns what fraction is the matix is split along be warned that it only pays attention to fully aliased situations so if only 2 of three levels appear the function will not flag them you have been warned.}
\usage{
what_frac(design_matrix)
}
\arguments{
\item{design_matrix}{This is the design matrix that is of interest. Each column is for the factors and each row for the treatments. Needs to be a matrix or data.frame.}
}
\value{
This will return a matrix with one column (entitiled Aliased with Intercept) and rows corresponding to the effects confounded with the intercept.
}
\description{
Alias Structure Identifier
Returns what fraction is the matix is split along be warned that it only pays attention to fully aliased situations so if only 2 of three levels appear the function will not flag them you have been warned.
}
